+++ core-ng/src/main/java/core/framework/impl/log/Appender.java
@@ -1,0 +1,10 @@
+package core.framework.impl.log;
+
+import core.framework.impl.log.filter.LogFilter;
+
+/**
+ * @author neo
+ */
+public interface Appender {
+    void append(ActionLog log, LogFilter filter);
+}