+++ core-ng/src/main/java/core/framework/impl/log/KafkaAppender.java
@@ -5,10 +5,8 @@ import core.framework.impl.kafka.ProducerMetrics;
 import core.framework.impl.log.filter.LogFilter;
 import core.framework.impl.log.message.ActionLogMessage;
 import core.framework.impl.log.message.LogTopics;
-import core.framework.impl.log.message.PerformanceStatMessage;
 import core.framework.impl.log.message.StatMessage;
 import core.framework.util.Maps;
-import core.framework.util.Network;
 import core.framework.util.Threads;
 import org.apache.kafka.clients.producer.Callback;
 import org.apache.kafka.clients.producer.KafkaProducer;
@@ -21,10 +19,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.time.Duration;
-import java.time.Instant;
-import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.UUID;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
@@ -33,21 +28,9 @@ import java.util.concurrent.atomic.AtomicBoolean;
 /**
  * @author neo
  */
-public final class KafkaAppender {
-    private static final int MAX_TRACE_LENGTH = 1000000; // 1M
-
-    public static KafkaAppender create(String uri, String appName) {
-        Map<String, Object> config = Maps.newHashMap();
-        config.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, uri);
-        config.put(ProducerConfig.ACKS_CONFIG, "0");    // no acknowledge to maximize performance
-        config.put(ProducerConfig.MAX_BLOCK_MS_CONFIG, Duration.ofSeconds(30).toMillis());  // metadata update timeout
-        config.put(ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder");
-        Producer<String, byte[]> producer = new KafkaProducer<>(config, new StringSerializer(), new ByteArraySerializer());
-        return new KafkaAppender(uri, appName, producer);
-    }
-
+public final class KafkaAppender implements Appender {
     public final ProducerMetrics producerMetrics;
-    final BlockingQueue<Object> queue = new LinkedBlockingQueue<>();
+    private final BlockingQueue<Object> queue = new LinkedBlockingQueue<>();
     private final Logger logger = LoggerFactory.getLogger(KafkaAppender.class);
     private final String appName;
     private final Producer<String, byte[]> producer;
@@ -63,9 +46,15 @@ public final class KafkaAppender {
         }
     };
 
-    KafkaAppender(String uri, String appName, Producer<String, byte[]> producer) {
+    public KafkaAppender(String uri, String appName) {
         this.appName = appName;
-        this.producer = producer;
+
+        Map<String, Object> config = Maps.newHashMap();
+        config.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, uri);
+        config.put(ProducerConfig.ACKS_CONFIG, "0");    // no acknowledge to maximize performance
+        config.put(ProducerConfig.MAX_BLOCK_MS_CONFIG, Duration.ofSeconds(30).toMillis());  // metadata update timeout
+        config.put(ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder");
+        producer = new KafkaProducer<>(config, new StringSerializer(), new ByteArraySerializer());
 
         producerMetrics = new ProducerMetrics("log-forwarder");
         producerMetrics.set(producer.metrics());
@@ -102,55 +91,12 @@ public final class KafkaAppender {
         producer.close(timeoutInMs <= 0 ? 1000 : timeoutInMs, TimeUnit.MILLISECONDS);
     }
 
-    void forward(ActionLog log, LogFilter filter) {
-        var message = new ActionLogMessage();
-        message.app = appName;
-        message.serverIP = Network.localHostAddress();
-        message.id = log.id;
-        message.date = log.date;
-        message.result = log.result();
-        message.refId = log.refId;
-        message.elapsed = log.elapsed;
-        message.cpuTime = log.cpuTime;
-        message.action = log.action;
-        message.errorCode = log.errorCode();
-        message.errorMessage = log.errorMessage;
-        message.context = log.context;
-        message.stats = log.stats;
-        Map<String, PerformanceStatMessage> performanceStats = new LinkedHashMap<>(log.performanceStats.size());
-        for (Map.Entry<String, PerformanceStat> entry : log.performanceStats.entrySet()) {
-            PerformanceStat stat = entry.getValue();
-            var statMessage = new PerformanceStatMessage();
-            statMessage.count = stat.count;
-            statMessage.totalElapsed = stat.elapsedTime;
-            statMessage.readEntries = stat.readEntries;
-            statMessage.writeEntries = stat.writeEntries;
-            performanceStats.put(entry.getKey(), statMessage);
-        }
-        message.performanceStats = performanceStats;
-        if (log.flushTraceLog()) {
-            var builder = new StringBuilder(log.events.size() << 8);  // length * 256 as rough initial capacity
-            for (LogEvent event : log.events) {
-                String traceMessage = event.logMessage(filter);
-                if (builder.length() + traceMessage.length() >= MAX_TRACE_LENGTH) {
-                    builder.append(traceMessage, 0, MAX_TRACE_LENGTH - builder.length());
-                    builder.append("...(truncated)");
-                    break;
-                }
-                builder.append(traceMessage);
-            }
-            message.traceLog = builder.toString();
-        }
-        queue.add(message);
+    @Override
+    public void append(ActionLog log, LogFilter filter) {
+        queue.add(MessageFactory.actionLog(log, appName, filter));
     }
 
     public void forward(Map<String, Double> stats) {
-        var message = new StatMessage();
-        message.id = UUID.randomUUID().toString();
-        message.date = Instant.now();
-        message.app = appName;
-        message.serverIP = Network.localHostAddress();
-        message.stats = stats;
-        queue.add(message);
+        queue.add(MessageFactory.stat(stats, appName));
     }
 }