+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
 ### 6.6.5 (8/15/2018 - )
+* log: make log-processor leverage built-in kafak support, to index action as well
 
 ### 6.6.4 (8/9/2018 - 8/15/2018)
 * kafka: update to 2.0.0